Extract logic to helper fn, add CW to PD reporter
