+++ .eslintrc.js
@@ -30,5 +30,6 @@ module.exports = {
   rules: {
     'require-jsdoc': 'off',
     '@typescript-eslint/explicit-module-boundary-types': 'off',
+    '@typescript-eslint/ban-ts-comment': 'off',
   },
 };
