+++ reporters/cloudwatch.ts
@@ -6,6 +6,11 @@ import path from 'path';
 
 import env from '../env.json';
 import { Logger } from '../src/utils/logger';
+import {
+  generateMessage,
+  getVideoName,
+  putMetric,
+} from '../src/utils/reporters';
 
 const suite = process.env.SUITE;
 const logDir = path.join(__dirname, '../logs');
@@ -16,22 +21,9 @@ const runIDFile = path.join(__dirname, `../${suite}.id.txt`);
 const uid = new Date().toISOString().substr(0, 16);
 const logger = new Logger({ logDir, logFile });
 
-module.exports = Pagerduty;
+module.exports = CloudWatch;
 
-function generateMessage(state: string, test: mocha.Test) {
-  return `${state} - ${test.titlePath().join(' - ')}`;
-}
-
-function getVideoName(parent: mocha.Suite): string {
-  if (parent.root && parent.file) {
-    const testFile = parent.file.split('/');
-    return testFile[testFile.length - 1]; // yields <filename>.ts
-  } else {
-    return getVideoName(<Mocha.Suite>parent.parent);
-  }
-}
-
-function Pagerduty(runner: mocha.Runner) {
+function CloudWatch(runner: mocha.Runner) {
   mocha.reporters.Base.call(this, runner);
   let passes = 0;
   let failures = 0;
@@ -66,7 +58,7 @@ function Pagerduty(runner: mocha.Runner) {
         testContext: test.titlePath()[0],
         testState: 'pending',
       });
-      await putMetric(test, 'pending');
+      await putMetric({suite, test, result: 'pending' });
     });
 
     runner.on('pass', async function (test) {
@@ -80,7 +72,7 @@ function Pagerduty(runner: mocha.Runner) {
         testContext: test.titlePath()[0],
         testState: 'pass',
       });
-      await putMetric(test, 'pass');
+      await putMetric({suite, test, result: 'pass' });
     });
 
     runner.on('fail', async function (test, err) {
@@ -99,7 +91,7 @@ function Pagerduty(runner: mocha.Runner) {
         error: err.message,
       });
 
-      await putMetric(test, 'fail', { video: video });
+      await putMetric({suite, test, result: 'fail', { video: video } });
     });
 
     runner.on('end', async function () {
@@ -119,48 +111,4 @@ function Pagerduty(runner: mocha.Runner) {
   }
 }
 
-async function putMetric(
-  test: mocha.Test,
-  result: string,
-  details: { [key: string]: string | number } = {}
-) {
-  const testContext = test.titlePath()[0];
-  const metricValue = result === 'fail' ? 1 : 0;
-  const credentials = env.isDev
-    ? new AWS.SharedIniFileCredentials({
-        profile: env.aws.profile,
-      })
-    : undefined;
-  const cw = await getCloudWatchClient(credentials);
-
-  // TODO: Make one big request at `runner.on('end')` with all results
-  const metric: PutMetricDataInput = {
-    MetricData: [
-      {
-        MetricName: 'Test Result',
-        Dimensions: [
-          { Name: 'uid', Value: `${suite}-${testContext}-${test.title}` },
-          { Name: 'suite', Value: suite ?? 'suite-missing' },
-          { Name: 'testName', Value: test.title },
-          { Name: 'testContext', Value: testContext },
-          { Name: 'testState', Value: result },
-        ],
-        Timestamp: new Date(),
-        Value: metricValue,
-      },
-    ],
-    Namespace: `editorial-tools-integration-tests`,
-  };
-
-  await cw.putMetricData(metric).promise();
-}
-
-export async function getCloudWatchClient(
-  credentials: AWS.SharedIniFileCredentials | undefined
-) {
-  return credentials
-    ? new AWS.CloudWatch({ credentials, region: 'eu-west-1' })
-    : new AWS.CloudWatch({ region: 'eu-west-1' });
-}
-
-mocha.utils.inherits(Pagerduty, mocha.reporters.Spec);
+mocha.utils.inherits(CloudWatch, mocha.reporters.Spec);
