+++ reporters/pagerduty.ts
@@ -1,11 +1,20 @@
 import mocha from 'mocha';
 import fs from 'fs';
 import path from 'path';
-import fetch from 'node-fetch';
 
 import { Logger } from '../src/utils/logger';
 import env from '../env.json';
+import {
+  generateMessage,
+  putMetric,
+  getVideoName,
+  callPagerduty,
+} from '../src/utils/reporters';
 const suite = process.env.SUITE;
+
+if (!suite) {
+  throw new Error('No suite passed into pagerduty.js as envar');
+}
 
 const logDir = path.join(__dirname, '../logs');
 const tmpDir = path.join(__dirname, '../tmp');
@@ -15,25 +24,12 @@ const runIDFile = `${tmpDir}/${suite}.id.txt`;
 // Yields `YYYY-DD-MMTHH-MM`
 const uid = new Date().toISOString().substr(0, 16);
 
-const routingKey = env.pagerduty.routingKey;
 const logger = new Logger({ logDir, logFile });
 
 module.exports = Pagerduty;
 
-function generateMessage(state: string, test: Mocha.Test) {
-  return `${state} - ${test.titlePath().join(' - ')}`;
-}
-
-function getVideoName(parent: Mocha.Suite): string {
-  if (parent.root && parent.file) {
-    const testFile = parent.file.split('/');
-    return testFile[testFile.length - 1]; // yields <filename>.ts
-  } else {
-    return getVideoName(parent.parent);
-  }
-}
-
 function Pagerduty(runner: Mocha.Runner) {
+  // @ts-ignore
   mocha.reporters.Base.call(this, runner);
   let passes = 0;
   let failures = 0;
@@ -68,6 +64,7 @@ function Pagerduty(runner: Mocha.Runner) {
         testContext: test.titlePath()[0],
         testState: 'pending',
       });
+      await putMetric({ suite, test, result: 'pending' });
       await callPagerduty(test, 'resolve');
     });
 
@@ -82,6 +79,7 @@ function Pagerduty(runner: Mocha.Runner) {
         testContext: test.titlePath()[0],
         testState: 'pass',
       });
+      await putMetric({ suite, test, result: 'pass' });
       await callPagerduty(test, 'resolve');
     });
 
@@ -103,7 +101,8 @@ function Pagerduty(runner: Mocha.Runner) {
         error: err.message,
       });
 
-      const video = getVideoName(test.parent);
+      const video = getVideoName(<Mocha.Suite>test.parent); // TODO: Think of a way to surface this information in CW so that we can correlate alarms, logs and metrics
+      await putMetric({ suite, test, result: 'fail' });
       await callPagerduty(test, 'trigger', {
         error: err.message,
         videosFolder: `https://s3.console.aws.amazon.com/s3/buckets/${env.videoBucket}/videos/${year}/${month}/${date}/?region=${region}&tab=overview`,
@@ -130,41 +129,5 @@ function Pagerduty(runner: Mocha.Runner) {
   }
 }
 
-async function callPagerduty(test: mocha.Test, action: string, details = {}) {
-  const url = 'https://events.pagerduty.com/v2/enqueue';
-
-  const data = {
-    routing_key: routingKey,
-    event_action: action,
-    dedup_key: test.title,
-    payload: {
-      summary: `${test.titlePath()[0]} - ${test.title}`,
-      source: 'https://github.com/guardian/editorial-tools-integration-tests',
-      severity: 'critical',
-      timestamp: new Date().toISOString(),
-      component: 'Editorial Tools Integration Tests',
-      links: 'https://gu.com',
-      custom_details: details,
-    },
-  };
-
-  const params = {
-    method: 'POST',
-    headers: { Accept: 'application/json', 'Content-Type': 'application/json' },
-    body: JSON.stringify(data),
-  };
-
-  const response = await fetch(url, params);
-  const json = await response.json();
-  if (!response.ok) {
-    console.error('PagerdutyReportError:', JSON.stringify(json));
-    logger.error({
-      message: `PagerdutyReportError: ${json.message}`,
-      error: json.errors,
-      data,
-      uid,
-    });
-  }
-}
-
+// @ts-ignore
 mocha.utils.inherits(Pagerduty, mocha.reporters.Spec);
